backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pyjwt
google-genai
pydantic
pydantic-settings
python-multipart
